requirements.txt
alembic==1.4.0
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
distlib==0.3.0
dominate==2.4.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-JWT==0.3.2
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.4.2
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
virtualenv==20.0.4
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
